kbnm/hostmanifest: Make write permission error different for darwin vs the rest
